Move action context to own file
